Reorganized file structure for Server to match Go
